requirements.txt
Django==4.2.6
django-dirtyfields==1.9.2
django-phonenumber-field==7.2.0
django-filter==23.3
django-redis==5.4.0
django-autoslug==1.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
six==1.16.0
Pillow==10.0.1
psycopg2-binary
phonenumbers==8.13.20
requests==2.31.0
PyJWT==2.8.0
celery==5.3.4
flower==2.0.1
drf-yasg==1.21.7
mkdocs==1.5.3









